Move AssemblyInfo reading out to separate reader
